Move webservice classes to different package
